+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.1.1] - 2026-07-05
+
+### Added
+
+- Support REGEX configuration for sensitive route
+- Added orig_cookie_vid field across all enforcer activities 
+
 ## [4.1.0] - 2026-01-29
 
 ### Added